requirements.txt
aniso8601==8.0.0
click==7.1.2
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.18
Werkzeug==1.0.1
matplotlib~=3.8.3
numpy~=1.26.4
netCDF4~=1.6.5
Cartopy~=0.22.0
shapely~=2.0.3
requests~=2.31.0
cftime~=1.6.3
pillow~=10.2.0
Pyrebase~=3.0.27